table fixes and additions to the gl table

+++ app/Database/Migrations/2023-04-12-015839_Factory.php
@@ -10,7 +10,7 @@ class Factory extends Migration
     {
         $this->forge->addField([
             'id' => [
-                'type' => 'INT',
+                'type' => 'BIGINT',
                 'unsigned' => true,
                 'auto_increment' => true
             ],
